fix(core): reuse sd-jwt disclosure code between vc and w3c v2 sd-jwt
